remove and renamed some folders
