+++ internal/resource/common/err/err.go
