+++ internal/storage/storage.go
@@ -13,5 +13,3 @@ var (
 
 	ErrAppointmentExists = errors.New("appointment exists")
 )
-
-// TODO add interface for ResourceNameStorage methods
